Fix typos after review.
Move $OFS='' to a more specific scope.  Add $ForcePoshGitPrompt param to posh-git.psm1 to force load of the posh-git prompt.

Moved git function to shared.ps1.  Eventually when I can figure how to remove this global function, I'd like this function to be add in BeforeAll and removed in AfterAll for the tests that actually need to mock Git.