Switch from SigStatus to SigType.
SignatureStatus appears to have four enum values that correspond to a signed file vs three for not signed.  Switching to check the SignatureType value as it will indicate Authenticode when file is Authenticode signed even if the hash doesn't match (been modified previously) or the cert is not trusted / incompatible.