Convert from untyped settings to strongly typed settings
This turns out to be important when you've changed the structure of the settings.  For instance, setting $GitPromptSettings.BranchColor = [consolecolor]::red doesn't work as expected because BranchColor is supposed to be a [PoshGitCellColor] object.  With the strongly typed settings, the above assignement will generate an error.  The assignment should be $GitPromptSettings.BranchColor.ForegroundColor = [ConsoleColor]::.